created jsonschemas for validation on routes that require it

+++ co-lab/backend/models/project.js
@@ -114,6 +114,10 @@ class Project {
       [id, username]
     );
     const deleted = qry.rows[0];
+    if (!deleted)
+      throw new BadRequestError(
+        `User with username ${username} not a cowriter on projectd with id ${id}`
+      );
     if (deleted.isOwner === true) {
       await db.query(`DELETE FROM projects WHERE id=$1`, [id]);
     }
